requirements.txt
streamlit==1.41.1
pymongo==4.10.1
anthropic==0.42.0
groq==0.18.0
requests==2.32.3
beautifulsoup4==4.12.3
python-dotenv==1.0.1
pandas==2.2.3
